Move Why Twenty under /resources and add Product to top nav
