+++ packages/twenty-website-new/src/app/[locale]/resources/why-twenty/editorial-four.data.ts
