+++ packages/twenty-website-new/src/app/[locale]/resources/why-twenty/editorial-one.data.ts
