+++ packages/twenty-website-new/src/app/[locale]/resources/why-twenty/editorial-three.data.ts
