+++ packages/twenty-website-new/src/app/[locale]/resources/why-twenty/hero.data.ts
