+++ packages/twenty-website-new/src/app/[locale]/resources/why-twenty/marquee.data.ts
