+++ packages/twenty-website-new/src/app/[locale]/resources/why-twenty/signoff.data.ts
