+++ packages/twenty-website-new/src/sections/Footer/data.ts
@@ -12,9 +12,14 @@ export const FOOTER_DATA: FooterDataType = {
       ctas: [],
       links: [
         { label: msg`Home`, href: '/', external: false },
+        { label: msg`Product`, href: '/product', external: false },
         { label: msg`Pricing`, href: '/pricing', external: false },
         { label: msg`Partners`, href: '/partners', external: false },
-        { label: msg`Why Twenty`, href: '/why-twenty', external: false },
+        {
+          label: msg`Why Twenty`,
+          href: '/resources/why-twenty',
+          external: false,
+        },
         { label: msg`Articles`, href: '/articles', external: false },
       ],
     },
